Move attributes and keys to respective files
Attributes and keybindings returned to their original location
with updated version.
