+++ tcell_compat.go
@@ -64,168 +64,6 @@ func tcellSize() (int, int) {
 	return screen.Size()
 }
 
-// Attribute affects the presentation of characters, such as color, boldness,
-// and so forth.
-type Attribute uint64
-
-const (
-	// ColorDefault is used to leave the Color unchanged from whatever system or teminal default may exist.
-	ColorDefault = Attribute(tcell.ColorDefault)
-
-	// AttrIsValidColor is used to indicate the color value is actually
-	// valid (initialized).  This is useful to permit the zero value
-	// to be treated as the default.
-	AttrIsValidColor = Attribute(tcell.ColorValid)
-
-	// AttrIsRGBColor is used to indicate that the Attribute value is RGB value of color.
-	// The lower order 3 bytes are RGB.
-	// (It's not a color in basic ANSI range 256).
-	AttrIsRGBColor = Attribute(tcell.ColorIsRGB)
-
-	// AttrColorBits is a mask where color is located in Attribute
-	AttrColorBits = 0xffffffffff // roughly 5 bytes, tcell uses 4 bytes and half-byte as a special flags for color (rest is reserved for future)
-
-	// AttrStyleBits is a mask where character attributes (e.g.: bold, italic, underline) are located in Attribute
-	AttrStyleBits = 0xffffff0000000000 // remaining 3 bytes in the 8 bytes Attribute (tcell is not using it, so we should be fine)
-)
-
-// Colors compatible with tcell colors
-const (
-	ColorBlack Attribute = AttrIsValidColor + iota
-	ColorRed
-	ColorGreen
-	ColorYellow
-	ColorBlue
-	ColorMagenta
-	ColorCyan
-	ColorWhite
-)
-
-// grayscale indexes (for backward compatibility with termbox-go original grayscale)
-var grayscale = []tcell.Color{
-	16, 232, 233, 234, 235, 236, 237, 238, 239, 240, 241, 242, 243, 244,
-	245, 246, 247, 248, 249, 250, 251, 252, 253, 254, 255, 231,
-}
-
-// Attributes are not colors, but affect the display of text.  They can
-// be combined.
-const (
-	AttrBold Attribute = 1 << (40 + iota)
-	AttrBlink
-	AttrReverse
-	AttrUnderline
-	AttrDim
-	AttrItalic
-	AttrNone Attribute = 0 // Just normal text.
-)
-
-// AttrAll is all the attributes turned on
-const AttrAll = AttrBold | AttrBlink | AttrReverse | AttrUnderline | AttrDim | AttrItalic
-
-// fixColor transform  Attribute into tcell.Color
-func fixColor(c Attribute) tcell.Color {
-	c = c & AttrColorBits
-	// Default color is 0 in tcell/v2 and was 0 in termbox-go, so we are good here
-	if c == ColorDefault {
-		return tcell.ColorDefault
-	}
-
-	tc := tcell.ColorDefault
-	// Check if we have valid color
-	if c&AttrIsValidColor != 0 {
-		tc = tcell.Color(c)
-	} else if c > 0 && c <= 256 {
-		// It's not valid color, but it has value in range 1-256
-		// This is old Attribute style of color from termbox-go (black=1, etc.)
-		// convert to tcell color (black=0|ColorValid)
-		tc = tcell.Color(c-1) | tcell.ColorValid
-	}
-
-	switch outMode {
-	case OutputTrue:
-		return tc
-	case OutputNormal:
-		tc &= tcell.Color(0xf) | tcell.ColorValid
-	case Output256:
-		tc &= tcell.Color(0xff) | tcell.ColorValid
-	case Output216:
-		tc &= tcell.Color(0xff)
-		if tc > 215 {
-			return tcell.ColorDefault
-		}
-		tc += tcell.Color(16) | tcell.ColorValid
-	case OutputGrayscale:
-		tc &= tcell.Color(0x1f)
-		if tc > 26 {
-			return tcell.ColorDefault
-		}
-		tc = grayscale[tc] | tcell.ColorValid
-	default:
-		return tcell.ColorDefault
-	}
-	return tc
-}
-
-func mkStyle(fg, bg Attribute) tcell.Style {
-	st := tcell.StyleDefault
-
-	// extract colors and attributes
-	if fg != ColorDefault {
-		st = st.Foreground(fixColor(fg))
-		st = setAttr(st, fg)
-	}
-	if bg != ColorDefault {
-		st = st.Background(fixColor(bg))
-		st = setAttr(st, bg)
-	}
-
-	return st
-}
-
-func setAttr(st tcell.Style, attr Attribute) tcell.Style {
-	if attr&AttrBold != 0 {
-		st = st.Bold(true)
-	}
-	if attr&AttrUnderline != 0 {
-		st = st.Underline(true)
-	}
-	if attr&AttrReverse != 0 {
-		st = st.Reverse(true)
-	}
-	if attr&AttrBlink != 0 {
-		st = st.Blink(true)
-	}
-	if attr&AttrDim != 0 {
-		st = st.Dim(true)
-	}
-	if attr&AttrItalic != 0 {
-		st = st.Italic(true)
-	}
-	return st
-}
-
-// GetColor creates a Color from a color name (W3C name). A hex value may
-// be supplied as a string in the format "#ffffff".
-func GetColor(color string) Attribute {
-	return Attribute(tcell.GetColor(color))
-}
-
-// Get256Color creates Attribute which stores ANSI color (0-255)
-func Get256Color(color int32) Attribute {
-	return Attribute(color) | AttrIsValidColor
-}
-
-// GetRGBColor creates Attribute which stores RGB color.
-// Color is passed as 24bit RGB value, where R << 16 | G << 8 | B
-func GetRGBColor(color int32) Attribute {
-	return Attribute(color) | AttrIsValidColor | AttrIsRGBColor
-}
-
-// NewRGBColor creates Attribute which stores RGB color.
-func NewRGBColor(r, g, b int32) Attribute {
-	return Attribute(tcell.NewRGBColor(r, g, b))
-}
-
 // tcellClear clears the screen with the given attributes.
 func tcellClear(fg, bg Attribute) {
 	st := mkStyle(fg, bg)
@@ -304,12 +142,6 @@ func tcellSetCell(x, y int, ch rune, fg, bg Attribute) {
 // EventType represents the type of event.
 type EventType uint8
 
-// Modifier represents the possible modifier keys.
-type Modifier tcell.ModMask
-
-// Key is a key press.
-type Key tcell.Key
-
 // Event represents an event like a key press, mouse action, or window resize.
 type Event struct {
 	Type   EventType
@@ -335,102 +167,9 @@ const (
 	EventRaw
 )
 
-// Keys codes.
-const (
-	KeyF1             = Key(tcell.KeyF1)
-	KeyF2             = Key(tcell.KeyF2)
-	KeyF3             = Key(tcell.KeyF3)
-	KeyF4             = Key(tcell.KeyF4)
-	KeyF5             = Key(tcell.KeyF5)
-	KeyF6             = Key(tcell.KeyF6)
-	KeyF7             = Key(tcell.KeyF7)
-	KeyF8             = Key(tcell.KeyF8)
-	KeyF9             = Key(tcell.KeyF9)
-	KeyF10            = Key(tcell.KeyF10)
-	KeyF11            = Key(tcell.KeyF11)
-	KeyF12            = Key(tcell.KeyF12)
-	KeyInsert         = Key(tcell.KeyInsert)
-	KeyDelete         = Key(tcell.KeyDelete)
-	KeyHome           = Key(tcell.KeyHome)
-	KeyEnd            = Key(tcell.KeyEnd)
-	KeyArrowUp        = Key(tcell.KeyUp)
-	KeyArrowDown      = Key(tcell.KeyDown)
-	KeyArrowRight     = Key(tcell.KeyRight)
-	KeyArrowLeft      = Key(tcell.KeyLeft)
-	KeyCtrlA          = Key(tcell.KeyCtrlA)
-	KeyCtrlB          = Key(tcell.KeyCtrlB)
-	KeyCtrlC          = Key(tcell.KeyCtrlC)
-	KeyCtrlD          = Key(tcell.KeyCtrlD)
-	KeyCtrlE          = Key(tcell.KeyCtrlE)
-	KeyCtrlF          = Key(tcell.KeyCtrlF)
-	KeyCtrlG          = Key(tcell.KeyCtrlG)
-	KeyCtrlH          = Key(tcell.KeyCtrlH)
-	KeyCtrlI          = Key(tcell.KeyCtrlI)
-	KeyCtrlJ          = Key(tcell.KeyCtrlJ)
-	KeyCtrlK          = Key(tcell.KeyCtrlK)
-	KeyCtrlL          = Key(tcell.KeyCtrlL)
-	KeyCtrlM          = Key(tcell.KeyCtrlM)
-	KeyCtrlN          = Key(tcell.KeyCtrlN)
-	KeyCtrlO          = Key(tcell.KeyCtrlO)
-	KeyCtrlP          = Key(tcell.KeyCtrlP)
-	KeyCtrlQ          = Key(tcell.KeyCtrlQ)
-	KeyCtrlR          = Key(tcell.KeyCtrlR)
-	KeyCtrlS          = Key(tcell.KeyCtrlS)
-	KeyCtrlT          = Key(tcell.KeyCtrlT)
-	KeyCtrlU          = Key(tcell.KeyCtrlU)
-	KeyCtrlV          = Key(tcell.KeyCtrlV)
-	KeyCtrlW          = Key(tcell.KeyCtrlW)
-	KeyCtrlX          = Key(tcell.KeyCtrlX)
-	KeyCtrlY          = Key(tcell.KeyCtrlY)
-	KeyCtrlZ          = Key(tcell.KeyCtrlZ)
-	KeyCtrlUnderscore = Key(tcell.KeyCtrlUnderscore)
-	KeyBackspace      = Key(tcell.KeyBackspace)
-	KeyBackspace2     = Key(tcell.KeyBackspace2)
-	KeyTab            = Key(tcell.KeyTab)
-	KeyEnter          = Key(tcell.KeyEnter)
-	KeyEsc            = Key(tcell.KeyEscape)
-	KeyPgdn           = Key(tcell.KeyPgDn)
-	KeyPgup           = Key(tcell.KeyPgUp)
-	KeyCtrlSpace      = Key(tcell.KeyCtrlSpace)
-	// KeySpace          = Key(tcell.Key(' '))
-	KeySpace = ' '
-	// KeyTilde = Key(tcell.Key('~'))
-	KeyTilde = '~'
-
-	// The following assignments were used in termbox implementation.
-	// In tcell, these are not keys per se. But in gocui we have them
-	// mapped to the keys so we have to use placeholder keys.
-	MouseLeft         = Key(tcell.KeyF63) // arbitrary assignments
-	MouseRight        = Key(tcell.KeyF62)
-	MouseMiddle       = Key(tcell.KeyF61)
-	MouseRelease      = Key(tcell.KeyF60)
-	MouseWheelUp      = Key(tcell.KeyF59)
-	MouseWheelDown    = Key(tcell.KeyF58)
-	MouseWheelLeft    = Key(tcell.KeyF57)
-	MouseWheelRight   = Key(tcell.KeyF56)
-	KeyCtrl2          = Key(tcell.KeyNUL) // termbox defines theses
-	KeyCtrl3          = Key(tcell.KeyEscape)
-	KeyCtrl4          = Key(tcell.KeyCtrlBackslash)
-	KeyCtrl5          = Key(tcell.KeyCtrlRightSq)
-	KeyCtrl6          = Key(tcell.KeyCtrlCarat)
-	KeyCtrl7          = Key(tcell.KeyCtrlUnderscore)
-	KeyCtrlSlash      = Key(tcell.KeyCtrlUnderscore)
-	KeyCtrlRsqBracket = Key(tcell.KeyCtrlRightSq)
-	KeyCtrlBackslash  = Key(tcell.KeyCtrlBackslash)
-	KeyCtrlLsqBracket = Key(tcell.KeyCtrlLeftSq)
-)
-
 var (
 	lastMouseKey tcell.ButtonMask = tcell.ButtonNone
 	lastMouseMod tcell.ModMask    = tcell.ModNone
-)
-
-// Modifiers.
-const (
-	ModAlt  = Modifier(tcell.ModAlt)
-	ModNone = Modifier(0)
-	// ModCtrl doesn't work with keyboard keys. Use CtrlKey in Key and ModNone. This is for mouse clicks only (tcell.v1)
-	// ModCtrl = Modifier(tcell.ModCtrl)
 )
 
 func makeEvent(tev tcell.Event) Event {
@@ -515,10 +254,4 @@ func makeEvent(tev tcell.Event) Event {
 	default:
 		return Event{Type: EventNone}
 	}
-}
-
-// tcellPollEvent blocks until an event is ready, and then returns it.
-func tcellPollEvent() Event {
-	ev := screen.PollEvent()
-	return makeEvent(ev)
 }
